+++ include/apprentice.h
@@ -6,11 +6,11 @@
 struct ApprenticeTrainer
 {
     u8 name[NUM_LANGUAGES - 1][PLAYER_NAME_LENGTH + 1]; // For all languages except the unused one.
-    u16 otId;
+    u32 otId;
     u8 facilityClass;
-    u16 species[APPRENTICE_SPECIES_COUNT];
+    u32 species[APPRENTICE_SPECIES_COUNT];
     u8 id;
-    u16 speechLost[EASY_CHAT_BATTLE_WORDS_COUNT];
+    u32 speechLost[EASY_CHAT_BATTLE_WORDS_COUNT];
 };
 
 extern const struct ApprenticeTrainer gApprentices[];
