+++ include/battle_factory.h
@@ -7,6 +7,6 @@ u8 GetFactoryMonFixedIV(u8 challengeNum, bool8 isLastBattle);
 void FillFactoryBrainParty(void);
 u8 GetNumPastRentalsRank(u8 battleMode, u8 lvlMode);
 u32 GetAiScriptsInBattleFactory(void);
-void SetMonMoveAvoidReturn(struct Pokemon *mon, u16 moveArg, u8 moveSlot);
+void SetMonMoveAvoidReturn(struct Pokemon *mon, u32 moveArg, u8 moveSlot);
 
 #endif // GUARD_BATTLE_FACTORY_H
