+++ include/battle_pyramid.h
@@ -2,7 +2,7 @@
 #define GUARD_BATTLE_PYRAMID_H
 
 void CallBattlePyramidFunction(void);
-u16 LocalIdToPyramidTrainerId(u8 localId);
+u32 LocalIdToPyramidTrainerId(u8 localId);
 bool8 GetBattlePyramidTrainerFlag(u8 eventId);
 void MarkApproachingPyramidTrainersAsBattled(void);
 void GenerateBattlePyramidWildMon(void);
@@ -11,14 +11,14 @@ u8 InBattlePyramid(void);
 bool8 InBattlePyramid_(void);
 void PausePyramidChallenge(void);
 void SoftResetInBattlePyramid(void);
-void CopyPyramidTrainerSpeechBefore(u16 trainerId);
-void CopyPyramidTrainerWinSpeech(u16 trainerId);
-void CopyPyramidTrainerLoseSpeech(u16 trainerId);
-u8 GetTrainerEncounterMusicIdInBattlePyramid(u16 trainerId);
-void GenerateBattlePyramidFloorLayout(u16 *mapArg, bool8 setPlayerPosition);
+void CopyPyramidTrainerSpeechBefore(u32 trainerId);
+void CopyPyramidTrainerWinSpeech(u32 trainerId);
+void CopyPyramidTrainerLoseSpeech(u32 trainerId);
+u8 GetTrainerEncounterMusicIdInBattlePyramid(u32 trainerId);
+void GenerateBattlePyramidFloorLayout(u32 *mapArg, bool8 setPlayerPosition);
 void LoadBattlePyramidObjectEventTemplates(void);
 void LoadBattlePyramidFloorObjectEventScripts(void);
 u8 GetNumBattlePyramidObjectEvents(void);
-u16 GetBattlePyramidPickupItemId(void);
+u32 GetBattlePyramidPickupItemId(void);
 
 #endif // GUARD_BATTLE_PYRAMID_H
